// feature-1.14.0/src/types/request.d.ts
export declare namespace RequestParams {
// ...
  interface GetPlan {
    create_by: Array<any>;
    name: string;
    product_id: Array<any>;
  }
  interface GetPlanListsTurnPage {
    create_by: string;
    name: string;
    page_index: number;
    page_size: number;
    product_id: number;
    status: number;
  }
// ...
}